Break out the `query` command
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`query` command.
