+++ pkg/cmd/pulumi/query/query.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2023, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package query
 
 import (
 	"context"
@@ -34,7 +34,7 @@ import (
 // intentionally disabling here for cleaner err declaration/assignment.
 //
 //nolint:vetshadow
-func newQueryCmd() *cobra.Command {
+func NewQueryCmd() *cobra.Command {
 	var stackName string
 
 	cmd := &cobra.Command{
